src/api/tox_requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-allauth==0.44.0
django-cors-headers==3.7.0
mysqlclient==2.0.1
pillow==7.1.0
factory-boy==3.2.0
pytest-django==3.10.0
flake8==3.7.9
six==1.15.0
